+++ src/main/java/org/c4marathon/assignment/Dto/AccountRequestDto.java
@@ -1,19 +1,17 @@
 package org.c4marathon.assignment.Dto;
 
-import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 public class AccountRequestDto {
-    private int balance;
-    private String type;
-    private int dailyWithdrawalLimit;
+	private int balance;
+	private String type;
+	private int dailyWithdrawalLimit;
 
-    @Builder
-    public AccountRequestDto(int balance, String type, int dailyWithdrawalLimit){
-        this.balance = balance;
-        this.type = type;
-        this.dailyWithdrawalLimit = dailyWithdrawalLimit;
-    }
+	public AccountRequestDto(int balance, String type, int dailyWithdrawalLimit) {
+		this.balance = balance;
+		this.type = type;
+		this.dailyWithdrawalLimit = dailyWithdrawalLimit;
+	}
 
 }
