+++ src/main/java/org/c4marathon/assignment/Dto/AccountResponseDto.java
@@ -5,17 +5,17 @@ import lombok.Getter;
 
 @Getter
 public class AccountResponseDto {
-    private Long accountId;
-    private int balance;
-    //일반계좌인지 적금계좌인지
-    private String type;
-    private int dailyWithdrawalLimit;
+	private Long accountId;
+	private int balance;
+	//일반계좌인지 적금계좌인지
+	private String type;
+	private int dailyWithdrawalLimit;
 
-    @Builder
-    public AccountResponseDto(Long accountId, int balance, String type, int dailyWithdrawalLimit){
-        this.accountId = accountId;
-        this.balance = balance;
-        this.type = type;
-        this.dailyWithdrawalLimit=dailyWithdrawalLimit;
-    }
+	@Builder
+	public AccountResponseDto(Long accountId, int balance, String type, int dailyWithdrawalLimit) {
+		this.accountId = accountId;
+		this.balance = balance;
+		this.type = type;
+		this.dailyWithdrawalLimit = dailyWithdrawalLimit;
+	}
 }
