+++ src/main/java/org/c4marathon/assignment/Dto/UserResponseDto.java
@@ -2,10 +2,8 @@ package org.c4marathon.assignment.Dto;
 
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
 public class UserResponseDto {
 	private Long userId;
 	private String name;
