+++ src/main/java/org/c4marathon/assignment/service/UserService.java
@@ -1,7 +1,5 @@
 package org.c4marathon.assignment.service;
 
-import java.time.LocalDate;
-
 import org.c4marathon.assignment.Dto.UserRequestDto;
 import org.c4marathon.assignment.Dto.UserResponseDto;
 import org.c4marathon.assignment.domain.Account;
@@ -46,44 +44,4 @@ public class UserService {
 			.registrationNum(savedUser.getRegistrationNum())
 			.build();
 	}
-
-	//적금 계좌 추가
-	@Transactional(isolation = Isolation.REPEATABLE_READ)
-	public void addSavingsAccount(Long userId, String type, int balance) {
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
-
-		//적금 계좌 생성
-		Account savingsAccount = new Account(type, balance, user);
-
-		//적금 계좌를 User 객체에 추가
-		user.addSavingAccount(savingsAccount);
-
-		//적금 계좌 저장
-		accountRepository.save(savingsAccount);
-	}
-
-	// 메인 계좌에서 적금 계좌로 송금
-	@Transactional(isolation = Isolation.REPEATABLE_READ)
-	public boolean transferToSavings(Long userId, Long savingsAccountId, int money) {
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
-
-		Account mainAccount = user.getMainAccount();
-		Account savingsAccount = accountRepository.findById(savingsAccountId)
-			.orElseThrow(() -> new IllegalArgumentException("적금 계좌를 찾을 수 없습니다."));
-
-		// 메인 계좌에서 출금
-		mainAccount.withdraw(money, LocalDate.now());
-
-		// 적금 계좌로 입금
-		savingsAccount.deposit(money);
-
-		// 계좌 저장
-		accountRepository.save(mainAccount);
-		accountRepository.save(savingsAccount);
-
-		return true;
-	}
-
 }
